+++ params/config_test.go
@@ -148,15 +148,15 @@ func TestMainnetBlobSchedule(t *testing.T) {
 	// Original EIP-4844 values
 	time := c.CancunTime.Uint64()
 	assert.Equal(t, uint64(6), c.GetMaxBlobsPerBlock(time))
-	assert.Equal(t, uint64(6*params.BlobGasPerBlob), c.GetMaxBlobGasPerBlock(time))
-	assert.Equal(t, uint64(3*params.BlobGasPerBlob), c.GetTargetBlobGasPerBlock(time))
+	assert.Equal(t, 6*params.BlobGasPerBlob, c.GetMaxBlobGasPerBlock(time))
+	assert.Equal(t, 3*params.BlobGasPerBlob, c.GetTargetBlobGasPerBlock(time))
 	assert.Equal(t, uint64(3338477), c.GetBlobGasPriceUpdateFraction(time))
 
 	// EIP-7691: Blob throughput increase
 	time = c.PragueTime.Uint64()
 	assert.Equal(t, uint64(9), c.GetMaxBlobsPerBlock(time))
-	assert.Equal(t, uint64(9*params.BlobGasPerBlob), c.GetMaxBlobGasPerBlock(time))
-	assert.Equal(t, uint64(6*params.BlobGasPerBlob), c.GetTargetBlobGasPerBlock(time))
+	assert.Equal(t, 9*params.BlobGasPerBlob, c.GetMaxBlobGasPerBlock(time))
+	assert.Equal(t, 6*params.BlobGasPerBlob, c.GetTargetBlobGasPerBlock(time))
 	assert.Equal(t, uint64(5007716), c.GetBlobGasPriceUpdateFraction(time))
 }
 
@@ -166,13 +166,13 @@ func TestGnosisBlobSchedule(t *testing.T) {
 	// Cancun values
 	time := c.CancunTime.Uint64()
 	assert.Equal(t, uint64(2), c.GetMaxBlobsPerBlock(time))
-	assert.Equal(t, uint64(2*params.BlobGasPerBlob), c.GetMaxBlobGasPerBlock(time))
-	assert.Equal(t, uint64(1*params.BlobGasPerBlob), c.GetTargetBlobGasPerBlock(time))
+	assert.Equal(t, 2*params.BlobGasPerBlob, c.GetMaxBlobGasPerBlock(time))
+	assert.Equal(t, 1*params.BlobGasPerBlob, c.GetTargetBlobGasPerBlock(time))
 	assert.Equal(t, uint64(1112826), c.GetBlobGasPriceUpdateFraction(time))
 
 	// should remain the same in Pectra for Gnosis
 	assert.Equal(t, uint64(2), c.GetMaxBlobsPerBlock(time))
-	assert.Equal(t, uint64(2*params.BlobGasPerBlob), c.GetMaxBlobGasPerBlock(time))
-	assert.Equal(t, uint64(1*params.BlobGasPerBlob), c.GetTargetBlobGasPerBlock(time))
+	assert.Equal(t, 2*params.BlobGasPerBlob, c.GetMaxBlobGasPerBlock(time))
+	assert.Equal(t, 1*params.BlobGasPerBlob, c.GetTargetBlobGasPerBlock(time))
 	assert.Equal(t, uint64(1112826), c.GetBlobGasPriceUpdateFraction(time))
 }
